langchain-zh/requirements.txt
# LangChain 核心库
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-anthropic>=0.1.0
langchain-community>=0.0.10

# 向量数据库
chromadb>=0.4.0
faiss-cpu>=1.7.0
pinecone-client>=2.2.0
weaviate-client>=3.17.0

# 数据库
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0

# 工具
python-dotenv>=1.0.0
requests>=2.31.0
duckduckgo-search>=4.0.0

# API 调用
httpx>=0.24.0

# 工具类
pydantic>=2.0.0
pydantic-settings>=2.0.0

# 数据处理
pandas>=2.0.0
numpy>=1.24.0

# 日志
loguru>=0.7.0

# 开发工具
pytest>=7.0.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0
isort>=5.12.0

# 文档生成
sphinx>=6.0.0
sphinx-rtd-theme>=1.2.0